change resolution type from short to int to reduce casting

+++ AppWebStream/Main.cpp
@@ -5,14 +5,14 @@
 #include "ScreenCapture.hpp"
 
 
-static HRESULT EncodeFrame (VideoEncoder & encoder, window_dc & wnd_dc, unsigned short dims[2]) {
+static HRESULT EncodeFrame (VideoEncoder & encoder, window_dc & wnd_dc, unsigned int dims[2]) {
     // create offscreen bitmap for screen capture (pad window size to be compatible with FFMPEG encoder)
     offscreen_bmp bmp(wnd_dc.dc, VideoEncoder::Align2(dims[0]), dims[1]);
 
     // copy window content encoder buffer
     auto * img_ptr = encoder.WriteFrameBegin();
     int scan_lines = bmp.CopyToRGBABuffer(wnd_dc.dc, (uint32_t*)img_ptr);
-    if (scan_lines != dims[1]) {
+    if (scan_lines != (int)dims[1]) {
         encoder.AbortWrite(); // still need to unlock buffer
         return E_FAIL;
     }
@@ -48,7 +48,7 @@ int main (int argc, char *argv[]) {
 
     // check window handle
     window_dc wnd_dc(win_handle);
-    unsigned short dims[2] = {static_cast<unsigned short>(wnd_dc.width()), static_cast<unsigned short>(wnd_dc.height())};
+    unsigned int dims[2] = {wnd_dc.width(), wnd_dc.height()};
     if (!wnd_dc.dc) {
         fprintf(stderr, "ERROR: Invalid window handle\n");
         return -1;
